pyproject.toml
[project]
name = "mae-data-gen"
version = "0.1.0"
description = "Traditional numerical solvers for generating Monge-Ampere equation reference data"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "scipy>=1.12",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
